+++ packages/plugin-mocker/src/runtimes/mock.ts
@@ -1,5 +1,4 @@
-import { getEnv } from '@zakijs/api'
-import { logger } from '@zakijs/api/lib/logger'
+import { getEnv, logger } from '@zakijs/api'
 import Attach from './attach'
 import { notNeedLog, notNeedNoMockWarn } from './constants'
 import MockItem from './mockItem'