+++ packages/runtime-mini/src/alipay/apisToAlipay.ts
@@ -1,7 +1,9 @@
-import { getGlobalObject } from '@zakijs/api/lib/env'
-import { logger } from '@zakijs/api/lib/logger'
-import type { IAPITransformConfig } from '@zakijs/api/lib/utils'
-import { transformApis } from '@zakijs/api/lib/utils'
+import {
+  getGlobalObject,
+  IAPITransformConfig,
+  logger,
+  transformApis
+} from '@zakijs/runtime-base'
 import { needPromisfiedApis } from './needPromisfiedApis'
 
 /**