+++ packages/runtime-mini/src/alipay/behaviorToAlipay.ts
@@ -1,4 +1,4 @@
-import { logger } from '@zakijs/api/lib/logger'
+import { logger } from '@zakijs/runtime-base'
 import type { BehaviorOptions, DefinitionFilter } from './utilsToAlipay'
 
 /**