+++ packages/core/CHANGELOG.md
@@ -3,12 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## 1.0.4-beta.0 (2023-03-08)
+## [1.0.6](https://github.com/eleme/morjs/compare/v1.0.5...v1.0.6) (2023-03-10)
+
+**Note:** Version bump only for package @morjs/core
 
 
-### Features
 
-* **packages:** reset all packages names ([e442c03](https://github.com/eleme/morjs/commit/e442c0375457c92ac0ee554f26cccf32f2bbf3c6))
+
+
+## [1.0.4](https://github.com/eleme/morjs/compare/v1.0.3...v1.0.4) (2023-03-09)
+
+
+### Bug Fixes
+
+* **core:** fix wechat npm build support ([32af3e2](https://github.com/eleme/morjs/commit/32af3e2f772909e46e8be84bc4f60820dd00d604))