+++ packages/create-mor/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.4-beta.2](https://github.com/eleme/morjs/compare/v1.0.4-beta.1...v1.0.4-beta.2) (2023-03-08)
+## [1.0.8](https://github.com/eleme/morjs/compare/v1.0.7...v1.0.8) (2023-03-15)
 
 **Note:** Version bump only for package create-mor
 
@@ -11,7 +11,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## [1.0.4-beta.1](https://github.com/eleme/morjs/compare/v1.0.4-beta.0...v1.0.4-beta.1) (2023-03-08)
+## [1.0.7](https://github.com/eleme/morjs/compare/v1.0.6...v1.0.7) (2023-03-10)
 
 **Note:** Version bump only for package create-mor
 
@@ -19,12 +19,25 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## 1.0.4-beta.0 (2023-03-08)
+## [1.0.6](https://github.com/eleme/morjs/compare/v1.0.5...v1.0.6) (2023-03-10)
+
+**Note:** Version bump only for package create-mor
 
 
-### Features
 
-* **packages:** reset all packages names ([e442c03](https://github.com/eleme/morjs/commit/e442c0375457c92ac0ee554f26cccf32f2bbf3c6))
+
+
+## [1.0.5](https://github.com/eleme/morjs/compare/v1.0.4...v1.0.5) (2023-03-10)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.4](https://github.com/eleme/morjs/compare/v1.0.3...v1.0.4) (2023-03-09)
+
+**Note:** Version bump only for package create-mor