+++ packages/runtime-base/CHANGELOG.md
@@ -3,6 +3,6 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 1.0.0 (2023-02-22)
+## [1.0.6](https://github.com/eleme/morjs/compare/v1.0.5...v1.0.6) (2023-03-10)
 
-**Note:** Version bump only for package @morjs/runtime-web
+**Note:** Version bump only for package @morjs/runtime-base