+++ packages/runtime-mini/CHANGELOG.md
@@ -3,12 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## 1.0.4-beta.0 (2023-03-08)
+## [1.0.6](https://github.com/eleme/morjs/compare/v1.0.5...v1.0.6) (2023-03-10)
 
 
-### Features
+### Bug Fixes
 
-* **packages:** reset all packages names ([e442c03](https://github.com/eleme/morjs/commit/e442c0375457c92ac0ee554f26cccf32f2bbf3c6))
+* **runtime-mini:** replace @morjs/api with @morjs/runtime-base to avoid cycle references ([84b4a0c](https://github.com/eleme/morjs/commit/84b4a0cef834e8bfabad4e2821b3f6414dd89ab9))
+
+
+
+
+
+## [1.0.4](https://github.com/eleme/morjs/compare/v1.0.3...v1.0.4) (2023-03-09)
+
+
+### Bug Fixes
+
+* **runtime-mini:** fix wechat npm build support ([096fb7c](https://github.com/eleme/morjs/commit/096fb7ce31547c71eac1e2c2f316a9f438c7da52))